Canvas上に描画完了
